Fix files larger than 2.15 GB can't be saved (fix #5935)
Before this fix the max file size without issues was 2.15GB,
now, Aseprite files can now exceed the current practical capacity
of hard drives.

Note: The on-disk header size field remains a uint32_t,
so it overflows for files > 4.30 GB. The only practical consequence
is that the progress bar becomes inaccurate when opening
files > 4.30 GB.

Also limit palette creation to 16384 colors when an RGBA sprite
contains a Black palette, preventing Aseprite from hanging due
to millions of unique colors in very large sprites.

+++ src/dio/aseprite_decoder.cpp
@@ -101,7 +101,7 @@ bool AsepriteDecoder::decode()
   // Read frame by frame to end-of-file
   for (frame_t frame = 0; frame < nframes; ++frame) {
     // Start frame position
-    const size_t frame_pos = tell();
+    const uint64_t frame_pos = tell();
     delegate()->progress((float)frame_pos / (float)header.size);
 
     // Read frame header
@@ -117,7 +117,7 @@ bool AsepriteDecoder::decode()
       // Read chunks
       for (uint32_t c = 0; c < frame_header.chunks; c++) {
         // Start chunk position
-        const size_t chunk_pos = tell();
+        const uint64_t chunk_pos = tell();
         delegate()->progress((float)chunk_pos / (float)header.size);
 
         // Read chunk information
@@ -307,7 +307,7 @@ bool AsepriteDecoder::decode()
 
 bool AsepriteDecoder::readHeader(AsepriteHeader* header)
 {
-  const size_t headerPos = tell();
+  const uint64_t headerPos = tell();
 
   header->size = read32();
   header->magic = read16();
@@ -1263,9 +1263,9 @@ Tileset* AsepriteDecoder::readTilesetChunk(Sprite* sprite,
 
   if (flags & ASE_TILESET_FLAG_EMBEDDED) {
     if (ntiles > 0) {
-      const size_t dataSize = read32(); // Size of compressed data
-      const size_t dataBeg = tell();
-      const size_t dataEnd = dataBeg + dataSize;
+      const uint32_t dataSize = read32(); // Size of compressed data
+      const uint64_t dataBeg = tell();
+      const uint64_t dataEnd = dataBeg + dataSize;
 
       base::buffer compressed;
       if (delegate()->cacheCompressedTilesets() && dataSize > 0) {
@@ -1457,7 +1457,7 @@ void AsepriteDecoder::readTilesData(Tileset* tileset, const AsepriteExternalFile
 {
   // Read as many user data chunks as tiles are in the tileset
   for (tile_index i = 0; i < tileset->size(); i++) {
-    const size_t chunk_pos = tell();
+    const uint64_t chunk_pos = tell();
     // Read chunk information
     const int chunk_size = read32();
     const int chunk_type = read16();
