+++ src/dio/file_interface.h
@@ -1,4 +1,5 @@
 // Aseprite Document IO Library
+// Copyright (c) 2026-present Igara Studio S.A.
 // Copyright (c) 2017-2018 David Capello
 //
 // This file is released under the terms of the MIT license.
@@ -22,10 +23,10 @@ public:
   virtual bool ok() const = 0;
 
   // Current position in the file
-  virtual size_t tell() = 0;
+  virtual uint64_t tell() = 0;
 
   // Jump to the given position in the file
-  virtual void seek(size_t absPos) = 0;
+  virtual void seek(uint64_t absPos) = 0;
 
   // Returns the next byte in the file or 0 if ok() = false
   virtual uint8_t read8() = 0;
@@ -40,8 +41,8 @@ class StdioFileInterface : public FileInterface {
 public:
   StdioFileInterface(FILE* file);
   bool ok() const override;
-  size_t tell() override;
-  void seek(size_t absPos) override;
+  uint64_t tell() override;
+  void seek(uint64_t absPos) override;
   uint8_t read8() override;
   size_t readBytes(uint8_t* buf, size_t n) override;
   void write8(uint8_t value) override;
